requirements.txt
fastapi
uvicorn
pydantic
python-jose
passlib
python-multipart
bcrypt
sqlalchemy
databases
pyyaml
aiosqlite